# docs/swagger.yaml
definitions:
  dao.ReqAddUser:
    properties:
      avatar:
        type: string
      createBy:
        type: string
      createdAt:
        type: string
      dataScope:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      nickname:
        type: string
      params:
        type: string
      password:
        type: string
      phone:
        type: string
      remark:
        type: string
      roleId:
        type: integer
      salt:
        type: string
      sex:
        type: string
      status:
        type: string
      updateBy:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      username:
        type: string
    required:
    - nickname
    - password
    - roleId
    - username
    type: object
...
  dao.ReqLoginUser:
    properties:
      password:
        example: "123"
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
...
  dao.ReqUpdateUser:
    properties:
      avatar:
        type: string
      createBy:
        type: string
      createdAt:
        type: string
      dataScope:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      nickname:
        type: string
      params:
        type: string
      password:
        type: string
      phone:
        type: string
      remark:
        type: string
      roleId:
        type: integer
      salt:
        type: string
      sex:
        type: string
      status:
        type: string
      updateBy:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
...
  response.Res:
    properties:
      code:
        type: integer
      data:
        type: object
      error:
        type: string
      msg:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /api/v1/login:
    post:
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dao.ReqLoginUser'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Res'
      summary: 用户登录
      tags:
      - User
...
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: 用户数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dao.ReqAddUser'
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      summary: 创建用户
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dao.ReqUpdateUser'
      responses:
        "200":
          description: '{"code": -1, "message": "修改失败"}'
          schema:
            type: string
      summary: 修改用户数据
      tags:
      - User
  /api/v1/user/{userId}:
    delete:
      description: 删除数据
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "删除失败"}'
          schema:
            type: string
      summary: 删除用户数据
      tags:
      - User
    get:
      description: 获取JSON
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      summary: 获取用户
      tags:
      - User
  /api/v1/user/profileAvatar:
    post:
      consumes:
      - multipart/form-data
      description: 获取JSON
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      summary: 修改头像
      tags:
      - User
  /api/v1/userlist:
    get:
      description: 获取JSON
      parameters:
      - description: username
        in: query
        name: username
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: 页条数
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      summary: 列表数据
      tags:
      - User
swagger: "2.0"